sail::adapter

Trait InboundDatagramRecvHalf

pub trait InboundDatagramRecvHalf:
    Sync
    + Send
    + Unpin {
    // Required method
    fn recv_from<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        buf: &'life1 mut [u8],
    ) -> Pin<Box<dyn Future<Output = ProxyResult<(usize, DatagramSource, SocksAddr)>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
}

The receive half.

Required Methods

fn recv_from<'life0, 'life1, 'async_trait>( &'life0 mut self, buf: &'life1 mut [u8], ) -> Pin<Box<dyn Future<Output = ProxyResult<(usize, DatagramSource, SocksAddr)>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Receives a single datagram message on the socket. On success, returns the number of bytes read, the source where this message originated and the destination this message shall be sent to.

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".