sail::adapter

Trait OutboundDatagramHandler

pub trait OutboundDatagramHandler:
    Send
    + Sync
    + Unpin {
    // Required methods
    fn connect_addr(&self) -> OutboundConnect;
    fn transport_type(&self) -> DatagramTransportType;
    fn handle<'a, 'async_trait>(
        &'a self,
        sess: &'a Session,
        transport: Option<AnyOutboundTransport>,
    ) -> Pin<Box<dyn Future<Output = Result<AnyOutboundDatagram>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'a: 'async_trait;
}

An outbound handler for outgoing UDP connections.

Required Methods

fn connect_addr(&self) -> OutboundConnect

Returns the address which the underlying transport should communicate with.

fn transport_type(&self) -> DatagramTransportType

Returns the transport type of this handler.

fn handle<'a, 'async_trait>( &'a self, sess: &'a Session, transport: Option<AnyOutboundTransport>, ) -> Pin<Box<dyn Future<Output = Result<AnyOutboundDatagram>> + Send + 'async_trait>>
where Self: 'async_trait, 'a: 'async_trait,

Handles a session with the transport. On success, returns an outbound datagram wraps the incoming transport.

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".