sail::adapter

Trait OutboundStreamHandler

pub trait OutboundStreamHandler:
    Send
    + Sync
    + Unpin {
    // Required methods
    fn connect_addr(&self) -> OutboundConnect;
    fn handle<'a, 'life0, 'async_trait>(
        &'a self,
        sess: &'a Session,
        lhs: Option<&'life0 mut AnyStream>,
        stream: Option<AnyStream>,
    ) -> Pin<Box<dyn Future<Output = Result<AnyStream>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'a: 'async_trait,
             'life0: 'async_trait;
}

An outbound handler for outgoing TCP conections.

Required Methods

fn connect_addr(&self) -> OutboundConnect

Returns the address which the underlying transport should communicate with.

fn handle<'a, 'life0, 'async_trait>( &'a self, sess: &'a Session, lhs: Option<&'life0 mut AnyStream>, stream: Option<AnyStream>, ) -> Pin<Box<dyn Future<Output = Result<AnyStream>> + Send + 'async_trait>>
where Self: 'async_trait, 'a: 'async_trait, 'life0: 'async_trait,

Handles a session with the given stream. On success, returns a stream wraps the incoming stream.

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".