sail::config

Module external_rule

Rule conditions that live in data files: GeoIP databases (.mmdb) and site lists (site.dat).

Structs

Mmdb

A country in a GeoIP database.

Enums

DomainKind

How a domain condition compares against a destination.

External

What an external rule adds to a rule.

Functions

geoip

code in the default GeoIP database.

geosite

The site group code in the default site list.

load

mmdb:<code>, mmdb:<file>:<code>, site:<code> or site:<file>:<code>. A relative file is looked up in the data directory.